refactor(eval): split engine.rs into module directory
